+++ resources/views/components/forms/form.blade.php
@@ -1,0 +1,11 @@
+@props([
+    'method' => 'POST',
+    'recaptcha' => true,
+])
+<form method="{{ $method }}" {{ $attributes }}>
+    @csrf
+    @if ($recaptcha)
+        {!! RecaptchaV3::field('store') !!}
+    @endif
+    {{ $slot }}
+</form>