+++ resources/views/components/forms/submit.blade.php
@@ -1,0 +1,3 @@
+<button type="submit" {{ $attributes->merge(['class' => 'btn btn-primary mt-3']) }}>
+    {{ strlen($slot) ? $slot : 'Odeslat' }}
+</button>