+++ routes/web.php
@@ -1,0 +1,5 @@
+<?php
+
+Route::get('/test-500', function (): never {
+    throw new Exception('Test 500 error');
+});