+++ src/Exceptions/TranslatorException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace InternetGuru\LaravelCommon\Exceptions;
+
+use Exception;
+
+class TranslatorException extends Exception
+{
+    //
+}